+++ test/samples/Alias.ts
@@ -1,0 +1,5 @@
+export type Alias = {
+	number: number
+	string: string
+	date: Date
+}